requirements.txt
futures==2.1.4
requests==1.2.3
sailthru-client==2.0.4